+++ Source/SwiftLintFramework/Extensions/Configuration+Remote.swift
@@ -119,7 +119,7 @@ internal extension Configuration.FileGraph.FilePath {
 
     private mutating func handleMissingNetwork(urlString: String, cachedFilePath: String?) throws -> String {
         if let cachedFilePath = cachedFilePath {
-            queuedPrint(
+            queuedPrintError(
                 "warning: No internet connectivity: Unable to load remote config from \"\(urlString)\". "
                     + "Using cached version as a fallback."
             )
@@ -141,11 +141,11 @@ internal extension Configuration.FileGraph.FilePath {
     ) throws -> String {
         if let cachedFilePath = cachedFilePath {
             if taskDone {
-                queuedPrint(
+                queuedPrintError(
                     "warning: Unable to load remote config from \"\(urlString)\". Using cached version as a fallback."
                 )
             } else {
-                queuedPrint(
+                queuedPrintError(
                     "warning: Timeout (\(timeout) sec): Unable to load remote config from \"\(urlString)\". "
                         + "Using cached version as a fallback."
                 )
@@ -170,7 +170,7 @@ internal extension Configuration.FileGraph.FilePath {
 
     private mutating func handleFileWriteFailure(urlString: String, cachedFilePath: String?) throws -> String {
         if let cachedFilePath = cachedFilePath {
-            queuedPrint("Unable to cache remote config from \"\(urlString)\". Using cached version as a fallback.")
+            queuedPrintError("Unable to cache remote config from \"\(urlString)\". Using cached version as a fallback.")
             self = .existing(path: cachedFilePath)
             return cachedFilePath
         } else {