+++ Source/swiftlint/Commands/Analyze.swift
@@ -43,6 +43,7 @@ extension SwiftLint {
                 benchmark: common.benchmark,
                 reporter: common.reporter,
                 quiet: quiet,
+                output: common.output,
                 cachePath: nil,
                 ignoreCache: true,
                 enableAllRules: false,